+++ CHANGELOG.md
@@ -14,9 +14,7 @@ Fleam release notes
 * Changes the Akka version to 2.6.8.
 * Changes usages of `ActorMaterializer` to `Materializer` and other smaller changes as per migration notes (https://doc.akka.io/docs/akka/current/project/migration-guide-2.5.x-2.6.x.html). 
 * Updated slf4j versions and overrides to eliminate inconsistency warnings.  
-* Adds the ability to choose the key used to do batch deletes with `SqsDelete`. If upgrading use `MessageId` as the
-  second parameter type to `toFlow` and `batched`. This allows you to use a custom id when desired instead of using the
-  default messageId which may be duplicated due to the at-least-once nature of SQS.
+* Changes `SqsDelete` to set the delete id to a sequential number instead of the message id. Previously duplicate messages failed. There will be no failure for duplicate messages now caused either by SQS or your own code.
 * Adds the option to process the SQS delete `BatchResult`s via a flow passed to the `SqsStreamDaemon`.
 * Changes `MissingMessageGroupId` to `MissingGroupingKey` for `SqsReduce` to make it more general.
 * Adds value class for `MessageGroupId`. Usages of SqsReduce will need to switch from `String` to `MessageGroupId`.